pyproject.toml
[project]
name = "bioplatform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "awsebcli>=3.23.2",
    "django>=4.2.20",
    "django-auth>=0.1.9",
    "pillow>=10.4.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.5",
]
